Move atoms from my model to atoms.tsx and move atoms.tsx into model layer.
